requirements.txt
typer==0.12.3
rich==13.7.1
tqdm==4.66.4
pypdf==4.2.0
openpyxl==3.1.5
pillow==10.3.0
pdf2image==1.17.0
pytesseract==0.3.10
python-dotenv==1.0.1
loguru==0.7.2